Refactor StringViewer to work well with all diagrams
